+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ansible: Add inventory README with deployment examples
 - Ansible: Add documentation for managing many domains via web panel
 
+- Sendry Web: Add DNS Check section to domain view page
+
 ### Fixed
 - Ansible: Fix package download URL for DEB and RPM packages